requirement.txt
beautifulsoup4==4.6.0
bleach==2.1.3
Django==2.0.3
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-mptt==0.9.0
html5lib==1.0.1
Pillow==5.0.0
pytz==2018.3
six==1.11.0
webencodings==0.5.1
